KakaoRedirect페이지 리팩토링 및  ts 전환
